fix(training-load): normalize activity training load to a TSS-like scale
The Activities table showed huge training-load numbers (e.g. 419,067)
because TrainingLoadCalculator returned the raw duration_seconds x avg_hr
product. The insight engine already normalizes the same product by
LOAD_NORMALIZER (5400) to a TSS-like scale (1h @ 150 bpm = 100); the
table was the only place showing the un-normalized value.

Apply the same 5400 normalizer in TrainingLoadCalculator and return a
Double (full precision), surfaced as Double on ActivityDTO. The frontend
formatter now renders 2 decimals. trainingLoad is derived at mapping
time (not stored), so no DB migration is needed.

+++ web/backend/athletedna/src/main/java/com/caznik/athletedna/infrastructure/web/dtos/ActivityDTO.java
@@ -10,6 +10,6 @@ public record ActivityDTO(
 	Integer avgHr,
 	Long externalStravaId,
 	Instant startDate,
-	Long trainingLoad,
+	Double trainingLoad,
 	String source
 ) {}
